Refactor file structure for tab components and home page migration
