+++ app/(tabs)/medicine/index.tsx
