+++ app/(tabs)/notes/index.tsx
