+++ app/(tabs)/settings/index.tsx
