+++ app/(tabs)/stats/index.tsx
